Port to Mutter to Clutter 1.0 API
- Use float instead of ClutterUnit
- clutter_actor_get_size/position() return floats (remove some
  stray usage of these functions in the default plugin as well.)
- Adapt to cogl_texture_new_from_data() changes
- Use blend strings to set up multitexturing
- Remove CLUTTER_UNITS_TO_FLOAT() usage

http://bugzilla.gnome.org/show_bug.cgi?id=585016

+++ src/compositor/mutter/compositor-mutter.c
@@ -577,7 +577,7 @@ mutter_window_set_property (GObject      *object,
           }
         else if (!newv && !priv->shadow && mutter_window_has_shadow (mw))
           {
-            guint        w, h;
+            gfloat       w, h;
             MetaDisplay *display = meta_screen_get_display (priv->screen);
             Mutter      *compositor;
